minor changes out of revision
